requirements.txt
# Python dependencies for Vercel API functions
numpy
scipy
pandas
scikit-learn
python-dotenv

